Check for missing branch labels on release-blockers
When an issue is labelled as a release-blocker, we generally also want it to have
branch-X labels applied as well identifying which branch or branches are affected
and thus should not release until it is resolved. This change adds a check whenever
labels are applied or removed that, if a 'release-blocker' label is present, at least
one branch-X label is as well, and, if not, comments asking the author to please apply
one.