Remove release-blocker comment from issue-open reply
The release-blocker vs branch label check is done any time an issue is labelled
or de-labelled, which includes when an issue is created i.e. the labelled event
is delivered on creation as well as the creation event.

This means the label checker will run on newly created issues independently
so we do not need to check release-blocker => branch label implication in
the creation hook and potentially mention it twice.

This done mean we may comment twice -- once about categorization and then
again about the release-blocker label -- on new issues, but release-blockers
are rare enough this should be a minor issue. Also, we really want someone to
see/react to a missing branch label so putting it in its own comment might be
a good idea anyway.